#7 change default checksum to sha1
On a pair of hot cache 1G files, the calculation time
as measured with checksum_speedtest.sh.

as a reference, the shell utilities from gnu coreutils
$time sha256sum *
and openssl command line
$time openssl sha256 *
were tested on the same input.

       rdfind  coreutils  openssl
md5    2.9     2.9        2.8
sha1   3.3     3.4        2.1
sha256 6.5     8.4        4.3

measured as elapsed time in seconds.
Based on this, it seems like using sha1 has an acceptable
performance cost but sha256 not (at least to use as a default)