Clarify that -makeresultsfile does not disable deletions
Just checking out this program today and I got the mistaken impression that -makeresultsfile was true by default and that as a result, nothing would be deleted. Boy was I wrong! I caught this before running it by reading the man file, but I almost missed it. I think a pointer to `-dryrun` would be helpful here.