fix buggy unit test, bug in check for getDigestLength() and make it into an assert
credit to github user @trollkarlen for finding this.

changed it to an assert since it should be caught in CI if it ever
happens.

while doing that, I detected that the unit test was not sufficient to
cover the assert.

the only consequence was that with SHA512 as checksum, an alarming output
was printed. there was no missed/false duplicate detection.